Made edits to the code. Removed the tests that are not required by the assignment
