+++ CarRentalPoint/CarRentalPoint/Entities/CarModel.cs
@@ -1,4 +1,6 @@
-﻿namespace CarRentalPoint.Entities;
+﻿using CarRentalPoint.Enums;
+
+namespace CarRentalPoint.Entities;
 
 /// <summary>
 /// Reference data for car models
@@ -7,6 +9,11 @@
 public class CarModel
 {
     /// <summary>
+    /// Unique identifier for the model
+    /// </summary>
+    public int Id { get; set; }
+
+    /// <summary>
     /// Name of the car model (e.g., "Toyota Camry", "BMW X5")
     /// </summary>
     public required string Name { get; set; }
@@ -14,7 +21,7 @@ public class CarModel
     /// <summary>
     /// Type of drive system (e.g., "Front-wheel", "All-wheel", "Rear-wheel")
     /// </summary>
-    public required string DriveType { get; set; }
+    public required CarDriveType DriveType { get; set; }
 
     /// <summary>
     /// Number of passenger seats in the vehicle
@@ -24,10 +31,10 @@ public class CarModel
     /// <summary>
     /// Body type of the vehicle (e.g., "Sedan", "SUV", "Hatchback")
     /// </summary>
-    public required string BodyType { get; set; }
+    public required CarBodyType BodyType { get; set; }
 
     /// <summary>
     /// Vehicle classification category (e.g., "A", "B", "C", "D", "E")
     /// </summary>
-    public required string CarClass { get; set; }
+    public required CarClass CarClass { get; set; }
 }
