+++ CarRentalPoint/CarRentalPoint/Enums/CarClass.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+namespace CarRentalPoint.Enums;
+
+/// <summary>
+/// Vehicle classification category according to European standards
+/// </summary>
+public enum CarClass
+{
+    /// <summary>
+    /// Mini cars
+    /// </summary>
+    A,
+    /// <summary>
+    /// Small cars
+    /// </summary>
+    B,
+    /// <summary>
+    /// Medium cars
+    /// </summary>
+    C,
+    /// <summary>
+    /// Large cars
+    /// </summary>
+    D,
+    /// <summary>
+    /// Executive cars
+    /// </summary>
+    E,
+    /// <summary>
+    /// Luxury cars
+    /// </summary>
+    F
+}
