+++ CarRentalPoint/CarRentalPoint/Enums/CarDriveType.cs
@@ -1,0 +1,26 @@
+﻿using System;
+namespace CarRentalPoint.Enums;
+
+/// <summary>
+/// Type of drive system
+/// </summary>
+public enum CarDriveType
+{
+    /// <summary>
+    /// Front-wheel drive
+    /// Power is delivered to the front wheels only
+    /// </summary>
+    FrontWheel,
+
+    /// <summary>
+    /// All-wheel drive (4WD) 
+    /// Power is distributed to all four wheels
+    /// </summary>
+    AllWheel,
+
+    /// <summary>
+    /// Rear-wheel drive 
+    /// Power is delivered to the rear wheels only
+    /// </summary>
+    RearWheel
+}
